project1/requirements/local.txt
# Local development requirements
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==9.11.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
